I have fixed the parts pointed out.
